create funtion moveS
